requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
streamlit>=1.32  # UI
requests>=2.31   
langgraph>=0.0.30
langchain-openai>=0.1.0 # Openai bases model 
langchain-community>=0.0.1   # if using langchain_community vectorstores
chromadb>=0.3.0              # Chroma backing store
python-dotenv>=1.0
pydantic>=2.0
langchain-groq
langchain-huggingface
sentence-transformers